Se ha reajustado toda la arquitectura para que el parseo dependa exclusivamente del flujo de parse_args. Esta, y todas las funciones auxiliares se han incluido en un script checker y lo que hacen es validar que los argumentos entran dentro de los rangos adecuados y, en caso de que no, devuelve un mensaje de error personalizado. Se han añadido varias funciones utils: ft_strlen, ft_strncmp... Falta hacer algún ajuste menor porque ahora mismo da fallo cuando no debería.

+++ philo/utils.c
@@ -6,13 +6,13 @@
 /*   By: jsanz-bo <jsanz-bo@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/08 23:51:34 by jsanz-bo          #+#    #+#             */
-/*   Updated: 2025/06/09 17:41:46 by jsanz-bo         ###   ########.fr       */
+/*   Updated: 2025/06/09 22:59:32 by jsanz-bo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philosophers.h"
 
-long long ft_atoi(const char *str)
+int ft_atoi(const char *str)
 {
 	char    sign;
 	int		r;
